Make BLE parser self-contained
