+++ ble_monitor/ble_parser
@@ -1,0 +1,1 @@
+../custom_components/ble_monitor/ble_parser
